+++ test/object-store/dictionary.cpp
@@ -18,6 +18,7 @@
 
 #include <catch2/catch.hpp>
 
+#include "collection_fixtures.hpp"
 #include "util/test_file.hpp"
 #include "util/index_helpers.hpp"
 
@@ -55,13 +56,24 @@ struct StringMaker<object_store::Dictionary> {
 };
 } // namespace Catch
 
-TEST_CASE("dictionary") {
+namespace cf = realm::collection_fixtures;
+
+TEMPLATE_TEST_CASE("dictionary types", "[dictionary]", cf::MixedVal, cf::Int, cf::Bool, cf::Float, cf::Double,
+                   cf::String, cf::Binary, cf::Date, cf::OID, cf::Decimal, cf::UUID, cf::BoxedOptional<cf::Int>,
+                   cf::BoxedOptional<cf::Bool>, cf::BoxedOptional<cf::Float>, cf::BoxedOptional<cf::Double>,
+                   cf::BoxedOptional<cf::OID>, cf::BoxedOptional<cf::UUID>, cf::UnboxedOptional<cf::String>,
+                   cf::UnboxedOptional<cf::Binary>, cf::UnboxedOptional<cf::Date>, cf::UnboxedOptional<cf::Decimal>)
+{
+    using T = typename TestType::Type;
+    using W = typename TestType::Wrapped;
+    using Boxed = typename TestType::Boxed;
+
     InMemoryTestFile config;
     config.cache = false;
     config.automatic_change_notifications = false;
     config.schema =
         Schema{{"object",
-                {{"value", PropertyType::Dictionary | PropertyType::String},
+                {{"value", PropertyType::Dictionary | TestType::property_type()},
                  {"links", PropertyType::Dictionary | PropertyType::Object | PropertyType::Nullable, "target"}}},
                {"target", {{"value", PropertyType::Int}}}};
 
@@ -88,15 +100,34 @@ TEST_CASE("dictionary") {
         REQUIRE(values_as_results.get_realm() == r);
     }
 
-    std::vector<std::string> keys = {"a", "b", "c"};
-    std::vector<std::string> values = {"apple", "banana", "clementine"};
-
+    auto values = TestType::values();
+    std::vector<std::string> keys;
     for (size_t i = 0; i < values.size(); ++i) {
-        dict.insert(keys[i], values[i]);
+        keys.push_back(util::format("key_%1", i));
     }
 
+    for (size_t i = 0; i < values.size(); ++i) {
+        dict.insert(keys[i], T(values[i]));
+    }
+
+    auto verify_keys_ordered = [&keys](Results& r) {
+        REQUIRE(r.size() == keys.size());
+        for (size_t i = 0; i < r.size(); ++i) {
+            REQUIRE(r.get<StringData>(i) == keys[i]);
+            REQUIRE(r.get_any(i) == keys[i]);
+        }
+    };
+
+    auto verify_values_ordered = [&values](Results& r) {
+        REQUIRE(r.size() == values.size());
+        for (size_t i = 0; i < values.size(); ++i) {
+            REQUIRE(r.get<T>(i) == values[i]);
+            REQUIRE(r.get_any(i) == Mixed(values[i]));
+        }
+    };
+
     SECTION("clear()") {
-        REQUIRE(dict.size() == 3);
+        REQUIRE(dict.size() == keys.size());
         values_as_results.clear();
         REQUIRE(dict.size() == 0);
         REQUIRE(values_as_results.size() == 0);
@@ -104,33 +135,35 @@ TEST_CASE("dictionary") {
 
     SECTION("get()") {
         for (size_t i = 0; i < values.size(); ++i) {
-            REQUIRE(dict.get<String>(keys[i]) == values[i]);
+            REQUIRE(dict.get<T>(keys[i]) == values[i]);
             auto val = dict.get(ctx, keys[i]);
-            REQUIRE(any_cast<std::string>(val) == values[i]);
+            REQUIRE(any_cast<Boxed>(val) == Boxed(values[i]));
         }
     }
 
     SECTION("insert()") {
         for (size_t i = 0; i < values.size(); ++i) {
             auto rev = values.size() - i - 1;
-            dict.insert(keys[i], values[rev]);
-            REQUIRE(dict.get<StringData>(keys[i]) == values[rev]);
+            dict.insert(keys[i], T(values[rev]));
+            REQUIRE(dict.get<T>(keys[i]) == values[rev]);
         }
         for (size_t i = 0; i < values.size(); ++i) {
-            dict.insert(ctx, keys[i], util::Any(values[i]));
-            REQUIRE(dict.get<StringData>(keys[i]) == values[i]);
+            dict.insert(ctx, keys[i], TestType::to_any(values[i]));
+            REQUIRE(dict.get<T>(keys[i]) == values[i]);
         }
     }
 
     SECTION("iteration") {
         for (size_t i = 0; i < values.size(); ++i) {
-            auto ndx = dict.find_any(values[i]);
+            auto ndx = dict.find_any(T(values[i]));
+            REQUIRE(ndx != realm::not_found);
             Dictionary::Iterator it = dict.begin() + ndx;
             REQUIRE((*it).first.get_string() == keys[i]);
-            REQUIRE((*it).second.get_string() == values[i]);
+            Mixed val_i{values[i]};
+            REQUIRE((*it).second == val_i);
             auto element = values_as_results.get_dictionary_element(ndx);
+            REQUIRE(element.second == val_i);
             REQUIRE(element.first == keys[i]);
-            REQUIRE(element.second.get_string() == values[i]);
             std::string key = keys_as_results.get<StringData>(ndx);
             REQUIRE(key == keys[i]);
             Mixed m = keys_as_results.get_any(ndx);
@@ -140,12 +173,53 @@ TEST_CASE("dictionary") {
 
     SECTION("keys sorted") {
         REQUIRE(keys_as_results.get_type() == PropertyType::String);
-        auto sorted = keys_as_results.sort({{"self", true}});
-        std::string key = sorted.get<StringData>(0);
-        REQUIRE(key == "a");
-        Mixed m = sorted.get_any(0);
-        REQUIRE(m.get_string() == "a");
-        REQUIRE_THROWS_WITH(sorted.get_any(4), "Requested index 4 greater than max 2");
+        SECTION("ascending") {
+            auto sorted = keys_as_results.sort({{"self", true}});
+            std::sort(begin(keys), end(keys), cf::less());
+            verify_keys_ordered(sorted);
+        }
+        SECTION("descending") {
+            auto sorted = keys_as_results.sort({{"self", false}});
+            std::sort(begin(keys), end(keys), cf::greater());
+            verify_keys_ordered(sorted);
+        }
+    }
+    SECTION("values sorted") {
+        REQUIRE(values_as_results.get_type() == TestType::property_type());
+        SECTION("ascending") {
+            auto sorted = values_as_results.sort({{"self", true}});
+            std::sort(begin(values), end(values), cf::less());
+            verify_values_ordered(sorted);
+        }
+        SECTION("descending") {
+            auto sorted = values_as_results.sort({{"self", false}});
+            std::sort(begin(values), end(values), cf::greater());
+            verify_values_ordered(sorted);
+        }
+    }
+
+    SECTION("keys distinct") {
+        REQUIRE(keys_as_results.get_type() == PropertyType::String);
+        keys.clear();
+        // set keys up in dictionary order
+        for (size_t i = 0; i < keys_as_results.size(); ++i) {
+            keys.push_back(std::string(keys_as_results.get<StringData>(i)));
+        }
+        auto distinct = keys_as_results.distinct({{"self"}});
+        verify_keys_ordered(distinct);
+    }
+
+    SECTION("values distinct") {
+        REQUIRE(values_as_results.get_type() == TestType::property_type());
+        SECTION("original") {
+            auto distinct = values_as_results.distinct({{"self"}});
+            values.clear();
+            // set values up in dictionary order
+            for (size_t i = 0; i < values_as_results.size(); ++i) {
+                values.push_back(values_as_results.get<T>(i));
+            }
+            verify_values_ordered(distinct);
+        }
     }
 
     SECTION("handover") {
@@ -155,8 +229,9 @@ TEST_CASE("dictionary") {
         REQUIRE(dict == dict2);
         ThreadSafeReference ref(values_as_results);
         auto results2 = ref.resolve<Results>(r).sort({{"self", true}});
+        std::sort(begin(values), end(values), cf::less());
         for (size_t i = 0; i < values.size(); ++i) {
-            REQUIRE(results2.get<String>(i) == values[i]);
+            REQUIRE(results2.get<T>(i) == values[i]);
         }
         r->begin_transaction();
         obj.remove();
@@ -190,47 +265,54 @@ TEST_CASE("dictionary") {
             // doesn't have dupes resulting in an unstable order
             advance_and_notify(*r);
             r->begin_transaction();
-            dict.erase("b");
+            dict.erase(keys[0]);
             r->commit_transaction();
 
             advance_and_notify(*r);
             r->begin_transaction();
-            dict.insert("d", "dade");
+            dict.insert(keys[0], T(values[0]));
             r->commit_transaction();
 
             advance_and_notify(*r);
-            auto ndx = values_as_results.index_of(StringData("dade"));
+            auto ndx = values_as_results.index_of(T(values[0]));
             REQUIRE_INDICES(change.insertions, ndx);
             REQUIRE_INDICES(rchange.insertions, ndx);
-            // "dade" ends up at the end of the sorted list
+            // values[0] is max(), so it ends up at the end of the sorted list
             REQUIRE_INDICES(srchange.insertions, values.size() - 1);
         }
 
         SECTION("replace value in dictionary") {
+            // Remove the existing copy of this value so that the sorted list
+            // doesn't have dupes resulting in an unstable order
             advance_and_notify(*r);
             r->begin_transaction();
-            dict.insert("b", "blueberry");
+            dict.erase(keys[0]);
             r->commit_transaction();
 
             advance_and_notify(*r);
-            auto ndx = values_as_results.index_of(StringData("blueberry"));
+            r->begin_transaction();
+            dict.insert(keys[1], T(values[0]));
+            r->commit_transaction();
+
+            advance_and_notify(*r);
+            auto ndx = values_as_results.index_of(T(values[0]));
             REQUIRE_INDICES(change.insertions);
             REQUIRE_INDICES(change.modifications, ndx);
             REQUIRE_INDICES(change.deletions);
         }
 
-        SECTION("remove value from list") {
+        SECTION("remove value from dictionary") {
             advance_and_notify(*r);
-            auto ndx = values_as_results.index_of(StringData("apple"));
+            auto ndx = values_as_results.index_of(T(values[0]));
+            auto ndx_sorted = sorted.index_of(T(values[0]));
             r->begin_transaction();
-            dict.erase("a");
+            dict.erase(keys[0]);
             r->commit_transaction();
 
             advance_and_notify(*r);
             REQUIRE_INDICES(change.deletions, ndx);
             REQUIRE_INDICES(rchange.deletions, ndx);
-            // apple comes first in the sorted result.
-            REQUIRE_INDICES(srchange.deletions, 0);
+            REQUIRE_INDICES(srchange.deletions, ndx_sorted);
         }
 
         SECTION("key based notification") {
@@ -242,11 +324,13 @@ TEST_CASE("dictionary") {
             advance_and_notify(*r);
 
             r->begin_transaction();
-            dict.insert("a", "apricot");
+            dict.insert(keys[0], T(values[1]));
             r->commit_transaction();
 
             advance_and_notify(*r);
-            REQUIRE(key_change.modifications[0].get_string() == "a");
+            REQUIRE(key_change.insertions.size() == 0);
+            REQUIRE(key_change.deletions.size() == 0);
+            REQUIRE(key_change.modifications[0].get_string() == keys[0]);
         }
 
         SECTION("clear list") {
@@ -361,7 +445,7 @@ TEST_CASE("dictionary") {
     }
 }
 
-TEST_CASE("embedded dictionary") {
+TEST_CASE("embedded dictionary", "[dictionary]") {
     InMemoryTestFile config;
     config.cache = false;
     config.automatic_change_notifications = false;